requirements.txt
flask
werkzeug
google-generativeai
pyttsx3
yt-dlp
